requirements.txt
beautifulsoup4==4.12.3
fitz==0.0.1.dev2
Flask==3.0.3
Flask_Cors==4.0.1
httpx==0.27.0
icrawler==0.6.9
lxml==5.2.2
matplotlib==3.7.5
numpy==1.24.4
openai==0.28.1
openai==0.28.0
opencv_python==4.9.0.80
pandas==2.0.3
pdf2image==1.17.0
Pillow==10.4.0
pydantic==1.10.15
pytesseract==0.3.10
python-dotenv==1.0.1
Requests==2.32.3
schedule==1.2.1
selenium==4.22.0
sentence_transformers==2.7.0
webdriver_manager==4.0.1
